refactor of rydantic models
